update the package names
